Rename Request.Body to RequestBody and Response.Body to ResponseBody.
Compelling alternative is ReadableBody and WritableBody, or BodySource and BodySink.
Which in theory permits reuse for serverside body makers. But I don't like those names
as much.
