+++ okcurl/src/main/java/com/squareup/okhttp/curl/Main.java
@@ -22,6 +22,7 @@ import com.squareup.okhttp.MediaType;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Protocol;
 import com.squareup.okhttp.Request;
+import com.squareup.okhttp.RequestBody;
 import com.squareup.okhttp.Response;
 import com.squareup.okhttp.internal.http.StatusLine;
 import com.squareup.okhttp.internal.spdy.Http20Draft12;
@@ -187,7 +188,7 @@ public class Main extends HelpOption implements Runnable {
     return "GET";
   }
 
-  private Request.Body getRequestBody() {
+  private RequestBody getRequestBody() {
     if (data == null) {
       return null;
     }
@@ -205,7 +206,7 @@ public class Main extends HelpOption implements Runnable {
       }
     }
 
-    return Request.Body.create(MediaType.parse(mimeType), bodyData);
+    return RequestBody.create(MediaType.parse(mimeType), bodyData);
   }
 
   Request createRequest() {
