+++ okhttp-tests/src/test/java/com/squareup/okhttp/internal/huc/JavaApiConverterTest.java
@@ -22,7 +22,9 @@ import com.squareup.okhttp.MediaType;
 import com.squareup.okhttp.OkHttpClient;
 import com.squareup.okhttp.Protocol;
 import com.squareup.okhttp.Request;
+import com.squareup.okhttp.RequestBody;
 import com.squareup.okhttp.Response;
+import com.squareup.okhttp.ResponseBody;
 import com.squareup.okhttp.internal.SslContextBuilder;
 import com.squareup.okhttp.internal.Util;
 import com.squareup.okhttp.mockwebserver.MockResponse;
@@ -432,7 +434,7 @@ public class JavaApiConverterTest {
   }
 
   @Test public void createJavaUrlConnection_responseHeadersOk() throws Exception {
-    Response.Body responseBody = createResponseBody("BodyText");
+    ResponseBody responseBody = createResponseBody("BodyText");
     Response okResponse = new Response.Builder()
         .request(createArbitraryOkRequest())
         .protocol(Protocol.HTTP_1_1)
@@ -609,7 +611,7 @@ public class JavaApiConverterTest {
             .url("http://insecure/request")
             .post(createRequestBody("RequestBody"))
             .build();
-    Response.Body responseBody = createResponseBody("ResponseBody");
+    ResponseBody responseBody = createResponseBody("ResponseBody");
     Response okResponse = createArbitraryOkResponse(okRequest).newBuilder()
         .protocol(Protocol.HTTP_1_1)
         .code(200)
@@ -633,7 +635,7 @@ public class JavaApiConverterTest {
             .url("https://secure/request")
             .post(createRequestBody("RequestBody") )
             .build();
-    Response.Body responseBody = createResponseBody("ResponseBody");
+    ResponseBody responseBody = createResponseBody("ResponseBody");
     Handshake handshake = Handshake.get("SecureCipher", Arrays.<Certificate>asList(SERVER_CERT),
         Arrays.<Certificate>asList(LOCAL_CERT));
     Response okResponse = createArbitraryOkResponse(okRequest).newBuilder()
@@ -770,14 +772,14 @@ public class JavaApiConverterTest {
     return createArbitraryOkResponse(createArbitraryOkRequest());
   }
 
-  private static Request.Body createRequestBody(String bodyText) {
-    return Request.Body.create(MediaType.parse("text/plain"), bodyText);
+  private static RequestBody createRequestBody(String bodyText) {
+    return RequestBody.create(MediaType.parse("text/plain"), bodyText);
   }
 
-  private static Response.Body createResponseBody(String bodyText) {
+  private static ResponseBody createResponseBody(String bodyText) {
     final Buffer source = new Buffer().writeUtf8(bodyText);
     final long contentLength = source.size();
-    return new Response.Body() {
+    return new ResponseBody() {
       @Override public MediaType contentType() {
         return MediaType.parse("text/plain; charset=utf-8");
       }
